Revert "Consolidate Depth Estimation Tasks (#807)"
This reverts commit 61608ff7eec4dbbe30739d6e87183485a4e88108.
