+++ src/lightly_train/_task_models/depth_estimation_components/convert_checkpoint_dav2.py
@@ -23,8 +23,11 @@ import torch
 from torch import Tensor
 
 from lightly_train._task_models import task_model_helpers
-from lightly_train._task_models.depth_estimation.task_model import (
-    DepthAnythingDepthEstimation,
+from lightly_train._task_models.dinov2_dav2_metric_depth_estimation.task_model import (
+    DepthAnythingV2MetricDepthEstimation,
+)
+from lightly_train._task_models.dinov2_dav2_relative_depth_estimation.task_model import (
+    DepthAnythingV2RelativeDepthEstimation,
 )
 
 logger = logging.getLogger(__name__)
@@ -69,7 +72,23 @@ _HF_WEIGHTS: dict[str, dict[str, str]] = {
     },
 }
 
-_SUPPORTED_MODELS: frozenset[str] = frozenset(_HF_WEIGHTS)
+# Task model class per (parsed) model name. The relative and metric models share the
+# official checkpoint layout and DPT shapes, so the same key remapping works for both.
+_MODEL_CLASSES: dict[
+    str,
+    type[DepthAnythingV2RelativeDepthEstimation]
+    | type[DepthAnythingV2MetricDepthEstimation],
+] = {
+    "dinov2/dav2-relative-small": DepthAnythingV2RelativeDepthEstimation,
+    "dinov2/dav2-relative-base": DepthAnythingV2RelativeDepthEstimation,
+    "dinov2/dav2-relative-large": DepthAnythingV2RelativeDepthEstimation,
+    "dinov2/dav2-metric-small-hypersim": DepthAnythingV2MetricDepthEstimation,
+    "dinov2/dav2-metric-base-hypersim": DepthAnythingV2MetricDepthEstimation,
+    "dinov2/dav2-metric-large-hypersim": DepthAnythingV2MetricDepthEstimation,
+    "dinov2/dav2-metric-small-vkitti": DepthAnythingV2MetricDepthEstimation,
+    "dinov2/dav2-metric-base-vkitti": DepthAnythingV2MetricDepthEstimation,
+    "dinov2/dav2-metric-large-vkitti": DepthAnythingV2MetricDepthEstimation,
+}
 
 # ``backbone.mask_token`` only exists for masked-image-modeling pretraining and is never
 # read during depth inference. The official checkpoint stores it with shape (1, 1, C)
@@ -122,7 +141,8 @@ def convert_checkpoint(
             "complying with its license terms."
         )
 
-    model = DepthAnythingDepthEstimation(
+    model_cls = _MODEL_CLASSES[parsed_name]
+    model = model_cls(
         model_name=parsed_name,
         load_weights=False,
     )
@@ -203,11 +223,11 @@ def main() -> None:
 
 def _parse_model_name(model_name: str) -> str:
     key = model_name.lower()
-    if key in _SUPPORTED_MODELS:
+    if key in _MODEL_CLASSES:
         return key
     raise ValueError(
         f"Model name '{model_name}' is not supported. Available models are: "
-        f"{sorted(_SUPPORTED_MODELS)}."
+        f"{sorted(_MODEL_CLASSES)}."
     )
 
 
@@ -224,7 +244,8 @@ def _download_huggingface_weights(repo_id: str, filename: str) -> Path:
 
 
 def _load_official_weights(
-    model: DepthAnythingDepthEstimation,
+    model: DepthAnythingV2RelativeDepthEstimation
+    | DepthAnythingV2MetricDepthEstimation,
     path: Path,
 ) -> None:
     state_dict = _load_state_dict_file(path)
