+++ src/lightly_train/_task_models/depth_estimation_components/convert_checkpoint_dav3.py
@@ -23,8 +23,11 @@ import torch
 from torch import Tensor
 
 from lightly_train._task_models import task_model_helpers
-from lightly_train._task_models.depth_estimation.task_model import (
-    DepthAnythingDepthEstimation,
+from lightly_train._task_models.dinov2_dav3_metric_depth_estimation.task_model import (
+    DepthAnythingV3MetricDepthEstimation,
+)
+from lightly_train._task_models.dinov2_dav3_relative_depth_estimation.task_model import (
+    DepthAnythingV3RelativeDepthEstimation,
 )
 
 logger = logging.getLogger(__name__)
@@ -41,7 +44,14 @@ _HF_WEIGHTS: dict[str, dict[str, str]] = {
     },
 }
 
-_SUPPORTED_MODELS: frozenset[str] = frozenset(_HF_WEIGHTS)
+_MODEL_CLASSES: dict[
+    str,
+    type[DepthAnythingV3MetricDepthEstimation]
+    | type[DepthAnythingV3RelativeDepthEstimation],
+] = {
+    "dinov2/dav3-relative-large": DepthAnythingV3RelativeDepthEstimation,
+    "dinov2/dav3-metric-large": DepthAnythingV3MetricDepthEstimation,
+}
 
 # ``backbone.mask_token`` only exists for masked-image-modeling pretraining and is
 # absent from the official inference checkpoints. It is never read during depth
@@ -69,7 +79,8 @@ def convert_checkpoint(
         The path to the exported checkpoint.
     """
     parsed_name = _parse_model_name(model_name)
-    model = DepthAnythingDepthEstimation(
+    model_cls = _MODEL_CLASSES[parsed_name]
+    model = model_cls(
         model_name=parsed_name,
         load_weights=False,
     )
@@ -143,11 +154,11 @@ def main() -> None:
 
 def _parse_model_name(model_name: str) -> str:
     key = model_name.lower()
-    if key in _SUPPORTED_MODELS:
+    if key in _MODEL_CLASSES:
         return key
     raise ValueError(
         f"Model name '{model_name}' is not supported. Available models are: "
-        f"{sorted(_SUPPORTED_MODELS)}."
+        f"{sorted(_MODEL_CLASSES)}."
     )
 
 
@@ -164,7 +175,8 @@ def _download_huggingface_weights(repo_id: str, filename: str) -> Path:
 
 
 def _load_official_weights(
-    model: DepthAnythingDepthEstimation,
+    model: DepthAnythingV3MetricDepthEstimation
+    | DepthAnythingV3RelativeDepthEstimation,
     path: Path,
 ) -> None:
     state_dict = _load_state_dict_file(path)
