+++ tests/_task_models/dinov2_dav2_metric_depth_estimation/__init__.py
